follow proper structure for setuptools
