+++ pyproject.toml
@@ -18,7 +18,7 @@ packages = [
 #]
 
 [tool.poetry.dependencies]
-python = "^3"
+python = ">=3.5"
 #click = "^7.1.2"
 Flask = "^1.1.2"
 #itsdangerous = "^1.1.0"
@@ -29,10 +29,10 @@ Flask = "^1.1.2"
 pywin32 = {version = "300", platform = "win32"}
 
 win10toast = {version = "^0.9", platform = "win32"}
-pydbus = { version = "^1.2.16", platform = "linux"}
+pydbus = { version = "^0.6", platform = "linux"}
 
 [tool.poetry.scripts]
-routEar = "routEar"
+routEar = "routEar.__main__:main"
 
 [build-system]
 requires = ["poetry>=0.12"]
